backend/requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug==2.3.7
flask-login==0.6.2
bcrypt==4.0.1
python-dateutil==2.8.2
flask-cors==4.0.0 